# najah-delivery-os/services/express-api/openapi.yaml
openapi: 3.0.3
info:
  title: Najah Delivery API
  description: Express API service for Najah Delivery OS - providing comprehensive delivery management endpoints
  version: 1.0.0
  contact:
    name: Najah Delivery Support
    email: support@najahdelivery.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.najahdelivery.com/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Merchant:
      type: object
      required:
        - name
        - email
        - phone
        - address
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        businessType:
          type: string
          enum: [restaurant, retail, pharmacy, grocery, other]
        address:
          type: object
          properties:
            ar:
              type: string
            en:
              type: string
            location:
              type: object
              properties:
                type:
                  type: string
                  enum: [Point]
                coordinates:
                  type: array
                  items:
                    type: number
                  minItems: 2
                  maxItems: 2
        status:
          type: string
          enum: [active, inactive, suspended]
...
    Courier:
      type: object
      required:
        - name
        - email
        - phone
        - vehicle
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        vehicle:
          type: object
          properties:
            type:
              type: string
              enum: [bicycle, motorcycle, car, van]
        status:
          type: string
          enum: [available, busy, offline, on_break]
        location:
          type: object
          properties:
            type:
              type: string
              enum: [Point]
            coordinates:
              type: array
              items:
                type: number
...
paths:
...
  /merchants:
    get:
      tags:
        - Merchants
      summary: List merchants
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive, suspended]
      responses:
        '200':
          description: List of merchants
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Merchant'
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  total:
                    type: integer

    post:
      tags:
        - Merchants
      summary: Create merchant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Merchant'
      responses:
        '201':
          description: Merchant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Merchant'
  /merchants/{id}:
    get:
      tags:
        - Merchants
      summary: Get merchant by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Merchant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Merchant'
        '404':
          description: Merchant not found
...
  /couriers:
    get:
      tags:
        - Couriers
      summary: List couriers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of couriers
          content:
            application/json:
              schema:
                type: object
                properties:
                  couriers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Courier'

    post:
      tags:
        - Couriers
      summary: Create courier
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Courier'
      responses:
        '201':
          description: Courier created
  /couriers/{id}/assign-route:
    post:
      tags:
        - Couriers
      summary: Assign route to courier
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                routePlanId:
                  type: string
      responses:
        '200':
          description: Route assigned
...